requirements.txt
python-dotenv
requests
openai
langchain
webdriver-manager
playwright
browser-use
streamlit
flask
waitress
langchain-core
langchain-community
duckduckgo-search
chromadb
pandas
numpy
ollama
langchain-chroma